specified 4 integers for course id and student id in tables

+++ db/db.go
@@ -24,7 +24,7 @@ func InitDB() {
 func createTables() {
 	createStudentsTable := `
 	CREATE TABLE IF NOT EXISTS students (
-		student_id INTEGER NOT NULL PRIMARY KEY AUTOINCREMENT,
+		student_id INTEGER(4) NOT NULL PRIMARY KEY AUTOINCREMENT,
 		first_name TEXT NOT NULL,
 		last_name TEXT NOT NULL,
 		email TEXT NOT NULL UNIQUE,
@@ -40,7 +40,7 @@ func createTables() {
 
 	createCoursesTable := `
 	CREATE TABLE IF NOT EXISTS courses (
-		course_id INTEGER NOT NULL PRIMARY KEY AUTOINCREMENT,
+		course_id INTEGER(4) NOT NULL PRIMARY KEY AUTOINCREMENT,
 		title VARCHAR(50) NOT NULL,
 		ticker VARCHAR(10) NOT NULL,
 		instructor_last_name VARCHAR(15) NOT NULL,
